+++ src/components/tasks/Checkbox.js
@@ -17,7 +17,7 @@ const CheckboxStyled = styled.input.attrs({ type: 'checkbox' })`
   &:checked {
     background-image: url(/images/checkmark.png);
     background-position: center;
-    animation: pulse 1s;
+    animation: pulse 0.75s;
     @keyframes pulse {
       0% {
         transform: scale(1);
