+++ src/index.js
@@ -3,9 +3,11 @@ import ReactDOM from 'react-dom'
 import App from './App'
 import * as serviceWorker from './serviceWorker'
 import { BrowserRouter as Router } from 'react-router-dom'
+import GlobalStyles from './GlobalStyles'
 
 ReactDOM.render(
   <React.StrictMode>
+    <GlobalStyles />
     <Router>
       <App />
     </Router>
